+++ packages/web3-eth/test/integration/subscription_logs.test.ts
@@ -17,9 +17,7 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 import WebSocketProvider from 'web3-providers-ws';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Contract, decodeEventABI } from 'web3-eth-contract';
-// eslint-disable-next-line import/no-extraneous-dependencies
-import { AbiEventFragment } from 'web3-eth-abi';
-import { Web3BaseProvider } from 'web3-types';
+import { AbiEventFragment, Web3BaseProvider } from 'web3-types';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import IpcProvider from 'web3-providers-ipc';
 import { Web3Eth } from '../../src';