+++ packages/web3/src/eth.exports.ts
@@ -1,4 +1,4 @@
-﻿/*
+/*
 This file is part of web3.js.
 
 web3.js is free software: you can redistribute it and/or modify
@@ -14,3 +14,12 @@ GNU Lesser General Public License for more details.
 You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
+
+export * from 'web3-eth';
+
+export * as abi from 'web3-eth-abi';
+export * as accounts from 'web3-eth-accounts';
+export * as contract from 'web3-eth-contract';
+export * as ens from 'web3-eth-ens';
+export * as personal from 'web3-eth-personal';
+export * as iban from 'web3-eth-iban';