+++ packages/web3-providers-ipc/CHANGELOG.md
@@ -67,3 +67,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   web3.js dependencies (#5757)
 
 ## [Unreleased]
+
+### Added
+
+-   Added named export for `IpcProvider` (#5771)